platform api 수정

+++ src/main/java/com/project/picktoon/controller/api/PlatformApiController.java
@@ -32,7 +32,10 @@ public class PlatformApiController {
     }
 
     @DeleteMapping("/{platformId}")
-    public void deletePlatform(@PathVariable("platformId") int id){
+    public ResponseEntity<Platform> deletePlatform(@PathVariable("platformId") int id){
+        if(platformService.getPlatformById(id)==null)
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
         platformService.deletePlatform(id);
+        return new ResponseEntity<>(HttpStatus.OK);
     }
 }
